+++ src/apps/DisciplineReleaseControl/Types/pipetest.ts
@@ -19,8 +19,10 @@ export type Pipetest = {
     overdue: string;
     dueDateTimePeriod: string;
     location: string;
-    lineNos: string[];
+    lines: Line[];
     mcPkgId: string;
+    hasDisconnectedEquipment: boolean;
+    hasIsolatedEquipment: boolean;
 };
 
 export type CheckList = {
@@ -78,6 +80,11 @@ export type InsulationBoxType = {
     objectStatus: string;
     object3dReference: string;
     procosysStatus: string;
+};
+
+export type Line = {
+    tagNo: string;
+    isCritical: boolean;
 };
 
 export type HTSidesheet = {