+++ src/apps/ReleaseControl/components/sidesheet/ReleaseControlSidesheet/Tabs/HistoryTab.tsx
@@ -2,12 +2,18 @@ import { useQuery } from 'react-query';
 import styled from 'styled-components';
 import { useReleaseControlContext } from '../../../../hooks/useReleaseControlContext';
 import { releaseControlQueries } from '../../../../queries/queries';
+import { Wrapper, WrapperFillerDiv } from '../sidesheetStyles';
 
 export const HistoryTab = (): JSX.Element => {
     const id = useReleaseControlContext(({ releaseControl }) => releaseControl.id);
     const { data: history } = useQuery(releaseControlQueries.historyQuery(id));
 
-    return <List>{history && history.map(({ id, title }) => <div key={id}>{title}</div>)}</List>;
+    return (
+        <Wrapper>
+            <List>{history && history.map(({ id, title }) => <div key={id}>{title}</div>)}</List>
+            <WrapperFillerDiv />
+        </Wrapper>
+    );
 };
 
 const List = styled.div`