+++ src/apps/ScopeChangeRequest/api/FAM/searchHtCable.ts
@@ -1,7 +1,7 @@
 import { generateExpressions, generateFamRequest } from '@equinor/fam-request-builder';
 import { httpClient } from '@equinor/lighthouse-portal-client';
 
-export async function searchHtCable(value: string, signal?: AbortSignal): Promise<any[]> {
+export async function searchHtCable(value: string): Promise<any[]> {
     const { FAM } = httpClient();
     const htExpression = generateExpressions('Register', 'Equals', ['HEAT_TRACING_CABLE']);
     const tagNoExpression = generateExpressions('TagNo', 'Equals', [value]);
@@ -23,9 +23,7 @@ export async function searchHtCable(value: string, signal?: AbortSignal): Promis
             'OpenWorkOrders',
             'Status',
             'InstalledCableLength',
-            'MountedOn',
             'TagHeated',
-            'TagMountedOn',
             'EstimatedCableLength',
             'SwitchBoardTagNos',
             'CircuitTagNos',
@@ -36,7 +34,6 @@ export async function searchHtCable(value: string, signal?: AbortSignal): Promis
     const res = await FAM.fetch('v0.1/dynamic/completion/custom_scope_tag/JCA', {
         body: JSON.stringify(request),
         method: 'POST',
-        signal,
     });
     return await res.json();
 }