+++ src/apps/ScopeChangeRequest/hooks/Search/useFAMSearch.ts
@@ -39,7 +39,15 @@ export function useFAMSearch(): FAMSearch {
                 );
             }
             case 'famtag': {
-                const items = await searchTag(searchValue, signal);
+                const items = await searchTag(searchValue);
+                items.map((tag: FamTag) => {
+                    tag.relatedHTCables =
+                        (tag.htCables !== null ? tag.htCables : '') +
+                        (tag.htCables !== null ? ', ' : '') +
+                        (tag.mountedOn_HTCables !== null ? tag.mountedOn_HTCables : '');
+                    tag.mountedOn = tag.mountedOnTagNo; //TODO - remove when backend fixed
+                    return tag;
+                });
                 return items.map(
                     (x: FamTag): TypedSelectOption => ({
                         label: `${x.tagNo}`,
@@ -51,7 +59,7 @@ export function useFAMSearch(): FAMSearch {
                 );
             }
             case 'htcable': {
-                const items = await searchHtCable(searchValue, signal);
+                const items = await searchHtCable(searchValue);
                 return items.map(
                     (x: FamTag): TypedSelectOption => ({
                         label: `${x.tagNo}`,