+++ src/packages/CircuitDiagram/src/types/pipetestTypes.ts
@@ -12,7 +12,7 @@ export type Pipetest = {
     overdue: string;
     dueDateTimePeriod: string;
     location: string;
-    lineNos: string[];
+    lines: Line[];
     mcPkgId: string;
 };
 
@@ -33,6 +33,11 @@ export type CheckList = {
     m03Forecast?: string;
     m04Actual?: string;
     underline?: string;
+};
+
+export type Line = {
+    tagNo: string;
+    isCritical: boolean;
 };
 
 export type Circuit = {