+++ src/packages/Utils/Hooks/index.ts
@@ -2,4 +2,4 @@ export * from './useCancellation';
 export * from './useElementData';
 export * from './useEventListener';
 export * from './useIsMounted';
-
+export * from './useClickOutside';